Add GNU Affero General Public License v3 to the project
This commit introduces the GNU Affero General Public License version 3, which ensures that users have the freedom to share, modify, and use the software while also requiring that the source code of modified versions be made available to users interacting with the software over a network. The license document is included in a new LICENSE.md file.